+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.0] - 2022-05-08
 ### Added
+- BoxCaster component
 - IEnable interface
 - BoxBodyGUI component
 - BoxBody component